tl;dr

+++ statistics_func/statistics.py
@@ -4,4 +4,4 @@ from coinranking import usd_exchange_rate, btc_exchange_price
 def price_calculate() -> float:
     usd_price = usd_exchange_rate()
     btc_price = btc_exchange_price()
-    return float(round(usd_price*btc_price, 2))
+    return round(usd_price*btc_price, 2)
